+++ src/TitleDetails/AddPieceModal/AddPieceModalContainer.test.js
@@ -50,25 +50,25 @@ describe('AddPieceModalContainer', () => {
 
   it('should display Edit Piece form', () => {
     const poLine = { id: 'poLineId', physical: { createInventory: 'None' }, locations: [{ locationId: '001' }] };
-    const initialValues = { caption: 'testcaption', format: 'Physical', id: 'id', poLineId: 'poLineId', titleId: 'titleId', locationId: '001' };
+    const initialValues = { enumeration: 'testenumeration', format: 'Physical', id: 'id', poLineId: 'poLineId', titleId: 'titleId', locationId: '001' };
     const locations = [{ name: 'Location', code: 'code', id: '001' }];
     const locationIds = ['001'];
 
     const { getByLabelText, getByText, queryByText } = renderAddPieceModalContainer(close, onSubmit, initialValues, 'instanceId', onCheckIn, poLine, locations, locationIds);
 
     // header is rendered
-    expect(getByText('ui-receiving.piece.caption')).toBeDefined();
-    expect(getByLabelText('ui-receiving.piece.caption')).toBeDefined();
+    expect(getByText('ui-receiving.piece.enumeration')).toBeDefined();
+    expect(getByLabelText('ui-receiving.piece.enumeration')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();
     expect(getByText('ui-receiving.piece.receiptDate')).toBeDefined();
     expect(queryByText('ui-receiving.piece.actions.quickReceive')).toBeTruthy();
-    fireEvent.input(getByLabelText('ui-receiving.piece.caption'));
+    fireEvent.input(getByLabelText('ui-receiving.piece.enumeration'));
   });
 
   it('should display Edit Received Piece form', () => {
     const poLine = { id: 'poLineId', physical: { createInventory: 'None' }, locations: [{ locationId: '001' }] };
     const piece = {
-      caption: 'testcaption',
+      enumeration: 'testenumeration',
       format: 'Physical',
       id: 'id',
       locationId: '001',
@@ -90,7 +90,7 @@ describe('AddPieceModalContainer', () => {
       locationIds,
     );
 
-    expect(getByLabelText('ui-receiving.piece.caption').disabled).toBeFalsy();
+    expect(getByLabelText('ui-receiving.piece.enumeration').disabled).toBeFalsy();
     expect(getByText('stripes-acq-components.piece.pieceFormat.physical')).toBeDefined();
     expect(getByLabelText('ui-receiving.piece.receiptDate').disabled).toBeFalsy();
     expect(queryByText('ui-receiving.piece.actions.quickReceive')).toBeFalsy();