+++ src/TitleDetails/ReceivedPiecesList/ReceivedPiecesList.test.js
@@ -7,7 +7,7 @@ import '@folio/stripes-acq-components/test/jest/__mock__';
 import ReceivedPiecesList from './ReceivedPiecesList';
 
 const pieces = [{
-  caption: 'ABA',
+  enumeration: 'v1',
   format: 'Electronic',
   receivedDate: '2020-02-06',
 }];
@@ -29,13 +29,13 @@ describe('Given Received Pieces List', () => {
 
     // header is rendered
     expect(getByText('ui-receiving.piece.barcode')).toBeDefined();
-    expect(getByText('ui-receiving.piece.caption')).toBeDefined();
+    expect(getByText('ui-receiving.piece.enumeration')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();
     expect(getByText('ui-receiving.piece.receivedDate')).toBeDefined();
     expect(getByText('ui-receiving.piece.request')).toBeDefined();
 
     // piece item is rendered
-    expect(getByText(pieces[0].caption)).toBeDefined();
+    expect(getByText(pieces[0].enumeration)).toBeDefined();
     expect(getByText('stripes-acq-components.piece.pieceFormat.electronic')).toBeDefined();
     expect(getByText(pieces[0].receivedDate)).toBeDefined();
   });