+++ src/common/utils/receive.js
@@ -32,7 +32,7 @@ export const checkInItems = (pieces, mutator) => {
       barcode: piece.barcode,
       callNumber: piece.callNumber,
       comment: piece.comment,
-      caption: piece.caption,
+      enumeration: piece.enumeration,
       supplement: piece.supplement,
       locationId: piece.locationId || null,
       holdingId: piece.holdingId || null,
@@ -55,7 +55,7 @@ export const checkInItems = (pieces, mutator) => {
           .filter(({ processingStatus }) => processingStatus.type === 'failure')
           .map((d) => ({
             ...d,
-            caption: pieces.find(({ id }) => id === d.pieceId)?.caption,
+            enumeration: pieces.find(({ id }) => id === d.pieceId)?.enumeration,
           }));
 
         return [...acc, ...errorResults];