+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * The Delete button is displayed in the Add piece pop up (new pice). Refs UIREC-146.
 * Ability to sort expected pieces by other table headers. Refs UIREC-147.
+* Update create, edit and receive piece forms with additional fields. Refs UIREC-135.
 
 ## [1.4.2](https://github.com/folio-org/ui-receiving/tree/v1.4.2) (2021-08-05)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.4.1...v1.4.2)